+++ src/Twig/Components/DashboardPerformanceChart.php
@@ -2,6 +2,7 @@
 
 namespace App\Twig\Components;
 
+use App\Repository\ContactMessageRepository;
 use App\Repository\ProjectRepository;
 use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 
@@ -9,14 +10,21 @@ use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 class DashboardPerformanceChart
 {
     private ProjectRepository $pr;
+    private ContactMessageRepository $cmr;
 
-    public function __construct(ProjectRepository $pr)
+    public function __construct(ProjectRepository $pr, ContactMessageRepository $cmr)
     {
         $this->pr = $pr;
+        $this->cmr = $cmr;
     }
 
-    public function getProjects(): int
+    public function getTotalProjectsAmount(): int
     {
-        return count($this->pr->getActiveAndShowInFrontendSortedByPosition());
+        return count($this->pr->findAll());
+    }
+
+    public function getTotalContactMessagesAmount(): int
+    {
+        return count($this->cmr->findAll());
     }
 }